Reworking PEP to prevent cross-contamination (#2394)
