+++ MetaMorpheus/EngineLayer/FdrAnalysis/FdrAnalysisEngine.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EngineLayer.CrosslinkSearch;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -275,18 +276,25 @@ namespace EngineLayer.FdrAnalysis
 
         public void Compute_PEPValue(FdrAnalysisResults myAnalysisResults, List<SpectralMatch> psms)
         {
-            if (psms[0].DigestionParams.Protease.Name == "top-down")
+            string searchType;
+            // Currently, searches of mixed data (bottom-up + top-down) are not supported
+            // PEP will be calculated based on the search type of the first file/PSM in the list, which isn't ideal
+            // This will be addressed in a future release
+            switch(psms[0].DigestionParams.Protease.Name)
             {
-                myAnalysisResults.BinarySearchTreeMetrics = PEP_Analysis_Cross_Validation.ComputePEPValuesForAllPSMsGeneric(psms, "top-down", this.FileSpecificParameters, this.OutputFolder);
+               case "top-down":
+                    searchType = "top-down";
+                    break;
+                default:
+                    searchType = "standard";
+                    break;
             }
-            else if (psms[0].DigestionParams.Protease.Name == "crosslink")
+            if (psms[0] is CrosslinkSpectralMatch)
             {
-                myAnalysisResults.BinarySearchTreeMetrics = PEP_Analysis_Cross_Validation.ComputePEPValuesForAllPSMsGeneric(psms, "crosslink", this.FileSpecificParameters, this.OutputFolder);
+                searchType = "crosslink";
             }
-            else
-            {
-                myAnalysisResults.BinarySearchTreeMetrics = PEP_Analysis_Cross_Validation.ComputePEPValuesForAllPSMsGeneric(psms, "standard", this.FileSpecificParameters, this.OutputFolder);
-            }
+            myAnalysisResults.BinarySearchTreeMetrics = new PepAnalysisEngine(psms, searchType, FileSpecificParameters, OutputFolder).ComputePEPValuesForAllPSMs();
+
         }
 
         /// <summary>
